+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -28,6 +28,7 @@ import net.geforcemods.securitycraft.blockentities.SonicSecuritySystemBlockEntit
 import net.geforcemods.securitycraft.blockentities.UsernameLoggerBlockEntity;
 import net.geforcemods.securitycraft.blocks.InventoryScannerFieldBlock;
 import net.geforcemods.securitycraft.blocks.LaserFieldBlock;
+import net.geforcemods.securitycraft.blocks.SecureRedstoneInterfaceBlock;
 import net.geforcemods.securitycraft.entity.camera.SecurityCamera;
 import net.geforcemods.securitycraft.entity.sentry.Sentry;
 import net.geforcemods.securitycraft.inventory.KeycardHolderMenu;
@@ -41,6 +42,7 @@ import net.geforcemods.securitycraft.items.SonicSecuritySystemItem;
 import net.geforcemods.securitycraft.misc.KeyBindings;
 import net.geforcemods.securitycraft.models.BlockMineModel;
 import net.geforcemods.securitycraft.models.DisguisableDynamicBakedModel;
+import net.geforcemods.securitycraft.models.SecureRedstoneInterfaceBakedModel;
 import net.geforcemods.securitycraft.particle.FloorTrapCloudParticle;
 import net.geforcemods.securitycraft.particle.InterfaceHighlightParticle;
 import net.geforcemods.securitycraft.renderers.BlockPocketManagerRenderer;
@@ -137,9 +139,12 @@ import net.minecraft.world.biome.BiomeColors;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.event.ColorHandlerEvent;
 import net.minecraftforge.client.event.ModelBakeEvent;
+import net.minecraftforge.client.event.ModelRegistryEvent;
 import net.minecraftforge.client.event.ParticleFactoryRegisterEvent;
 import net.minecraftforge.client.event.TextureStitchEvent;
 import net.minecraftforge.client.model.ModelDataManager;
+import net.minecraftforge.client.model.ModelLoader;
+import net.minecraftforge.client.model.data.IDynamicBakedModel;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.RegistryObject;
 import net.minecraftforge.fml.client.registry.ClientRegistry;
@@ -181,7 +186,8 @@ public class ClientHandler {
 			SCContent.SCANNER_DOOR.get(),
 			SCContent.SCANNER_TRAPDOOR.get(),
 			SCContent.SECURITY_CAMERA.get(),
-			SCContent.SECURE_REDSTONE_INTERFACE.get(),
+			//Excluded because it has its own custom baked model
+			//SCContent.SECURE_REDSTONE_INTERFACE.get(),
 			SCContent.SENTRY_DISGUISE.get(),
 			SCContent.SONIC_SECURITY_SYSTEM.get(),
 			SCContent.TROPHY_SYSTEM.get(),
@@ -191,6 +197,12 @@ public class ClientHandler {
 	public static final ResourceLocation LINKING_STATE_PROPERTY = new ResourceLocation(SecurityCraft.MODID, "linking_state");
 
 	private ClientHandler() {}
+
+	@SubscribeEvent
+	public static void onModelRegistry(ModelRegistryEvent event) {
+		ModelLoader.addSpecialModel(new ResourceLocation(SecurityCraft.MODID, "block/secure_redstone_interface_sender_on"));
+		ModelLoader.addSpecialModel(new ResourceLocation(SecurityCraft.MODID, "block/secure_redstone_interface_receiver_on"));
+	}
 
 	@SubscribeEvent
 	public static void onModelBake(ModelBakeEvent event) {
@@ -216,14 +228,24 @@ public class ClientHandler {
 				"stone"
 		};
 		//@formatter:on
-
 		Map<ResourceLocation, IBakedModel> modelRegistry = event.getModelRegistry();
+		Block sri = SCContent.SECURE_REDSTONE_INTERFACE.get();
+		IBakedModel poweredSriSender = modelRegistry.get(new ResourceLocation(SecurityCraft.MODID, "block/secure_redstone_interface_sender_on"));
+		IBakedModel poweredSriReceiver = modelRegistry.get(new ResourceLocation(SecurityCraft.MODID, "block/secure_redstone_interface_receiver_on"));
 
 		for (Block block : disguisableBlocks.get()) {
 			for (BlockState state : block.getStateDefinition().getPossibleStates()) {
-				registerDisguisedModel(modelRegistry, block.getRegistryName(), state.getValues().entrySet().stream().map(StateHolder.PROPERTY_ENTRY_TO_STRING_FUNCTION).collect(Collectors.joining(",")));
+				registerDisguisedModel(modelRegistry, state, DisguisableDynamicBakedModel::new);
 			}
 		}
+
+		for (BlockState state : sri.getStateDefinition().getPossibleStates()) {
+			if (state.getValue(SecureRedstoneInterfaceBlock.SENDER))
+				registerDisguisedModel(modelRegistry, state, oldModel -> new SecureRedstoneInterfaceBakedModel(poweredSriSender, oldModel));
+			else
+				registerDisguisedModel(modelRegistry, state, oldModel -> new SecureRedstoneInterfaceBakedModel(poweredSriReceiver, oldModel));
+		}
+
 		for (String mine : mines) {
 			registerBlockMineModel(event, new ResourceLocation(SecurityCraft.MODID, mine.replace("_ore", "") + "_mine"), new ResourceLocation(mine));
 		}
@@ -231,10 +253,12 @@ public class ClientHandler {
 		registerBlockMineModel(event, new ResourceLocation(SecurityCraft.MODID, "quartz_mine"), new ResourceLocation("nether_quartz_ore"));
 	}
 
-	private static void registerDisguisedModel(Map<ResourceLocation, IBakedModel> modelRegistry, ResourceLocation rl, String stateString) {
+	private static void registerDisguisedModel(Map<ResourceLocation, IBakedModel> modelRegistry, BlockState state, Function<IBakedModel, IDynamicBakedModel> modelFunction) {
+		ResourceLocation rl = state.getBlock().getRegistryName();
+		String stateString = state.getValues().entrySet().stream().map(StateHolder.PROPERTY_ENTRY_TO_STRING_FUNCTION).collect(Collectors.joining(","));
 		ModelResourceLocation mrl = new ModelResourceLocation(rl, stateString);
 
-		modelRegistry.put(mrl, new DisguisableDynamicBakedModel(modelRegistry.get(mrl)));
+		modelRegistry.put(mrl, modelFunction.apply(modelRegistry.get(mrl)));
 	}
 
 	private static void registerBlockMineModel(ModelBakeEvent event, ResourceLocation mineRl, ResourceLocation realBlockRl) {